power-libperfmgr: Implement basic screen on/off logic
Set SUSTAINED_PERFORMANCE as the default hint whenever the screen
turns on, and deactivate it appropriately when the screen goes off.
Subsequently it should be possible to keep hints cached across
screen state transitions, without mixing up the need for actual
purpose-bound hints (like the camera app boosting) and the sustained one.
This way we can guarantee that the correct hint is in use.